sfeed_curses: write out plumbercmd argument in a verbose manner
Write it in a more verbose, clear and C90-style manner.

This workarounds a bug in scc too (reported upstream and will be fixed of
course). All tools in sfeed can now be compiled with minicurses-mode with the
scc compiler and musl libc without any modifications needed.

scc: http://www.simple-cc.org/