sfeed_curses: cleanup code for handling read items from an URL file
This makes it easier to reuse the code from reading plain-text list of items
from a file and uses a struct urls to keep its state in a cleaner way.

+ some small code-style improvements and reshuffling.

Inspired by a question from NRK by mail about implementing a "read later"
feature.

(A quick local experiment shows it can be done in ~70 lines of added code).