atom, json, mbox: fix reading past the buffer with an escaped NUL byte (\ NUL)
This would skip checking the end of the string of checking a NUL byte, because
the iteration was done before checking it.

It would proceed into the data that comes after. Note that sfeed itself can't
generate such malformed data itself.

Example input:

	0	title	link	content\	html

Would incorrect print "contenthtml" as the content.