sfeedrc.5: document each feed name should be unique
Asked via e-mail. This confused a person which converted an OPML file which
yielded an sfeedrc with duplicate names or multiple "unnamed" entries.

This caused sfeed_update to fail, because it tried to merge files with the same
name and causing a race-condition because the temporary file with the same name
was already moved.